rename: api 공통 파일 (fetchAPI, path) utils 디렉터리로 이동
