+++ src/apis/utils/fetchAPI.ts
@@ -1,4 +1,4 @@
-import APIError from "../error/APIError";
+import APIError from "../../error/APIError";
 
 const BASE_URL = "https://api.themoviedb.org/3";
 const API_KEY = process.env.TMDB_ACCESS_KEY;
